# Performance[MQB,BMQ]: return constructed blobs as pointers