--- a/src/groups/bmq/bmqa/bmqa_message.t.cpp
+++ b/src/groups/bmq/bmqa/bmqa_message.t.cpp
@@ -166,6 +166,10 @@ static void test2_validPushMessagePrint()
     bdlbb::PooledBlobBufferFactory bufferFactory(
         4 * 1024,
         bmqtst::TestHelperUtil::allocator());
+    bmqp::BlobPoolUtil::BlobSpPool blobSpPool(
+        bmqp::BlobPoolUtil::createBlobPool(
+            &bufferFactory,
+            bmqtst::TestHelperUtil::allocator()));
     bmqa::Event                    event;
 
     EventImplSp& implPtr = reinterpret_cast<EventImplSp&>(event);
@@ -188,7 +192,7 @@ static void test2_validPushMessagePrint()
               bsl::strlen(buffer));
 
     // Create PushEventBuilder
-    bmqp::PushEventBuilder peb(&bufferFactory,
+    bmqp::PushEventBuilder peb(&blobSpPool,
                                bmqtst::TestHelperUtil::allocator());
     ASSERT_EQ(sizeof(bmqp::EventHeader), static_cast<size_t>(peb.eventSize()));
     ASSERT_EQ(sizeof(bmqp::EventHeader),
@@ -250,6 +254,10 @@ static void test3_messageProperties()
     bdlbb::PooledBlobBufferFactory bufferFactory(
         4 * 1024,
         bmqtst::TestHelperUtil::allocator());
+    bmqp::BlobPoolUtil::BlobSpPool blobSpPool(
+        bmqp::BlobPoolUtil::createBlobPool(
+            &bufferFactory,
+            bmqtst::TestHelperUtil::allocator()));
 
     const int               queueId = 4321;
     const bmqt::MessageGUID guid;
@@ -280,7 +288,7 @@ static void test3_messageProperties()
     bdlbb::BlobUtil::append(&payload, buffer, bsl::strlen(buffer));
 
     // Create PushEventBuilder
-    bmqp::PushEventBuilder peb(&bufferFactory,
+    bmqp::PushEventBuilder peb(&blobSpPool,
                                bmqtst::TestHelperUtil::allocator());
 
     // Add SubQueueInfo option
@@ -422,6 +430,10 @@ static void test4_subscriptionHandle()
     bdlbb::PooledBlobBufferFactory bufferFactory(
         4 * 1024,
         bmqtst::TestHelperUtil::allocator());
+    bmqp::BlobPoolUtil::BlobSpPool blobSpPool(
+        bmqp::BlobPoolUtil::createBlobPool(
+            &bufferFactory,
+            bmqtst::TestHelperUtil::allocator()));
     bdlbb::Blob payload(&bufferFactory, bmqtst::TestHelperUtil::allocator());
 
     queueSp->setId(queueId);
@@ -445,7 +457,7 @@ static void test4_subscriptionHandle()
             bmqtst::TestHelperUtil::allocator());
 
         // Create PushEventBuilder
-        bmqp::PushEventBuilder peb(&bufferFactory,
+        bmqp::PushEventBuilder peb(&blobSpPool,
                                    bmqtst::TestHelperUtil::allocator());
         ASSERT_EQ(0, peb.messageCount());
 
@@ -510,7 +522,7 @@ static void test4_subscriptionHandle()
             bmqtst::TestHelperUtil::allocator());
 
         // Create PushEventBuilder
-        bmqp::PushEventBuilder peb(&bufferFactory,
+        bmqp::PushEventBuilder peb(&blobSpPool,
                                    bmqtst::TestHelperUtil::allocator());
         ASSERT_EQ(0, peb.messageCount());
 
@@ -558,7 +570,7 @@ static void test4_subscriptionHandle()
             bmqtst::TestHelperUtil::allocator());
 
         // Create PutEventBuilder
-        bmqp::PutEventBuilder builder(&bufferFactory,
+        bmqp::PutEventBuilder builder(&blobSpPool,
                                       bmqtst::TestHelperUtil::allocator());
         ASSERT_EQ(0, builder.messageCount());
 
@@ -599,7 +611,7 @@ static void test4_subscriptionHandle()
             bmqtst::TestHelperUtil::allocator());
 
         // Create AckEventBuilder
-        bmqp::AckEventBuilder builder(&bufferFactory,
+        bmqp::AckEventBuilder builder(&blobSpPool,
                                       bmqtst::TestHelperUtil::allocator());
         ASSERT_EQ(0, builder.messageCount());