--- a/src/groups/bmq/bmqa/bmqa_mocksession.cpp
+++ b/src/groups/bmq/bmqa/bmqa_mocksession.cpp
@@ -298,7 +298,12 @@ Event MockSessionUtil::createAckEvent(const bsl::vector<AckParams>& acks,
     implPtr = EventImplSp(new (*alloc) bmqimp::Event(g_bufferFactory_p, alloc),
                           alloc);
 
-    bmqp::AckEventBuilder ackBuilder(bufferFactory, alloc);
+    // TODO: deprecate `createAckEvent` with bufferFactory arg and introduce
+    // another function with BlobSpPool arg.
+    bmqa::Session::BlobSpPool blobSpPool(
+        bmqp::BlobPoolUtil::createBlobPool(bufferFactory, allocator));
+
+    bmqp::AckEventBuilder ackBuilder(&blobSpPool, alloc);
     for (size_t i = 0; i != acks.size(); ++i) {
         const AckParams&   params   = acks[i];
         const QueueImplSp& impQueue = reinterpret_cast<const QueueImplSp&>(
@@ -336,7 +341,12 @@ Event MockSessionUtil::createPushEvent(
     implPtr = EventImplSp(new (*alloc) bmqimp::Event(g_bufferFactory_p, alloc),
                           alloc);
 
-    bmqp::PushEventBuilder pushBuilder(bufferFactory, alloc);
+    // TODO: deprecate `createPushEvent` with bufferFactory arg and introduce
+    // another function with BlobSpPool arg.
+    bmqa::Session::BlobSpPool blobSpPool(
+        bmqp::BlobPoolUtil::createBlobPool(bufferFactory, allocator));
+
+    bmqp::PushEventBuilder pushBuilder(&blobSpPool, alloc);
 
     for (size_t i = 0; i != pushEventParams.size(); ++i) {
         const QueueImplSp& queueImplPtr = reinterpret_cast<const QueueImplSp&>(
@@ -946,6 +956,8 @@ int MockSession::start(const bsls::TimeInterval& timeout)
 MockSession::MockSession(const bmqt::SessionOptions& options,
                          bslma::Allocator*           allocator)
 : d_blobBufferFactory(1024, allocator)
+, d_blobSpPool(
+      bmqp::BlobPoolUtil::createBlobPool(&d_blobBufferFactory, allocator))
 , d_eventHandler_mp(0)
 , d_calls(allocator)
 , d_eventsAndJobs(allocator)
@@ -990,6 +1002,8 @@ MockSession::MockSession(bslma::ManagedPtr<SessionEventHandler> eventHandler,
                          const bmqt::SessionOptions&            options,
                          bslma::Allocator*                      allocator)
 : d_blobBufferFactory(1024, allocator)
+, d_blobSpPool(
+      bmqp::BlobPoolUtil::createBlobPool(&d_blobBufferFactory, allocator))
 , d_eventHandler_mp(eventHandler)
 , d_calls(bslma::Default::allocator(allocator))
 , d_eventsAndJobs(bslma::Default::allocator(allocator))
@@ -1514,7 +1528,7 @@ void MockSession::loadMessageEventBuilder(MessageEventBuilder* builder)
                                  g_bufferFactory_p,
                                  d_allocator_p);
 
-    eventImplSpRef->configureAsMessageEvent(&d_blobBufferFactory);
+    eventImplSpRef->configureAsMessageEvent(&d_blobSpPool);
     eventImplSpRef->setMessageCorrelationIdContainer(
         d_corrIdContainer_sp.get());
 }
@@ -1538,7 +1552,7 @@ void MockSession::loadConfirmEventBuilder(ConfirmEventBuilder* builder)
     }
 
     new (builderImplRef.d_buffer.buffer())
-        bmqp::ConfirmEventBuilder(&d_blobBufferFactory, d_allocator_p);
+        bmqp::ConfirmEventBuilder(&d_blobSpPool, d_allocator_p);
 
     builderImplRef.d_builder_p = reinterpret_cast<bmqp::ConfirmEventBuilder*>(
         builderImplRef.d_buffer.buffer());