--- a/src/groups/bmq/bmqimp/bmqimp_application.cpp
+++ b/src/groups/bmq/bmqimp/bmqimp_application.cpp
@@ -553,6 +553,9 @@ Application::Application(
 , d_channelsTip(&d_allocator)
 , d_blobBufferFactory(sessionOptions.blobBufferSize(),
                       d_allocators.get("BlobBufferFactory"))
+, d_blobSpPool(
+      bmqp::BlobPoolUtil::createBlobPool(&d_blobBufferFactory,
+                                         d_allocators.get("BlobSpPool")))
 , d_scheduler(bsls::SystemClockType::e_MONOTONIC, &d_allocator)
 , d_channelFactory(ntcCreateInterfaceConfig(sessionOptions, allocator),
                    &d_blobBufferFactory,
@@ -582,11 +585,13 @@ Application::Application(
                                      negotiationMessage,
                                      sessionOptions.connectTimeout(),
                                      &d_blobBufferFactory,
+                                     &d_blobSpPool,
                                      allocator),
       allocator)
 , d_connectHandle_mp()
 , d_brokerSession(&d_scheduler,
                   &d_blobBufferFactory,
+                  &d_blobSpPool,
                   d_sessionOptions,
                   eventHandlerCB,
                   bdlf::MemFnUtil::memFn(&Application::stateCb, this),