--- a/src/groups/bmq/bmqimp/bmqimp_application.cpp
+++ b/src/groups/bmq/bmqimp/bmqimp_application.cpp
@@ -584,7 +584,6 @@ Application::Application(
       NegotiatedChannelFactoryConfig(&d_statChannelFactory,
                                      negotiationMessage,
                                      sessionOptions.connectTimeout(),
-                                     &d_blobBufferFactory,
                                      &d_blobSpPool,
                                      allocator),
       allocator)